// src/openapi/model/getPatrolSelfTableOfAutolivCommand.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PatrolSelfType } from './patrolSelfType';


export interface GetPatrolSelfTableOfAutolivCommand { 
    product_id?: string;
    machine_code?: string | null;
    patrol_self_type?: PatrolSelfType;
}
export namespace GetPatrolSelfTableOfAutolivCommand {
}